+++ src/components/useState/UseState.jsx
@@ -1,0 +1,27 @@
+import { useState } from 'react';
+import './UseState.css';
+
+function App() {
+  let [counter, setCounter] = useState(15);
+  const addValue = () => {
+    if (counter < 20) {
+      setCounter(counter + 1);
+    }
+  };
+  const removeValue = () => {
+    if (counter > 0) {
+      setCounter(counter - 1);
+    }
+  };
+
+  return (
+    <>
+      <h4>Start up with react project | Aman started builing the React App </h4>
+      <h2>Counter Value : {counter}</h2>
+      <button onClick={addValue}>Add Value</button>
+      <button onClick={removeValue}>Remove Value</button>
+    </>
+  );
+}
+
+export default App;
